+++ src/app/Balloon.App.Idp/.container.config.php
@@ -7,30 +7,57 @@ use Balloon\App\Idp\Storage\MongoDB;
 use OAuth2\Server as OAuth2Server;
 use OAuth2\OpenID\GrantType\AuthorizationCode;
 use OAuth2\GrantType\UserCredentials;
+use OAuth2\Storage\UserCredentialsInterface;
+use OAuth2\Storage\RefreshTokenInterface;
 use OAuth2\GrantType\RefreshToken;
 use Micro\Auth\Auth;
 use Balloon\App\Idp\Auth\Token;
+use Balloon\App\Idp\GrantType\UserCredentialsMultiFactor;
+use Balloon\App\Idp\Storage\UserCredentials as UserCredentialsStorage;
+use Balloon\App\Idp\Storage\UserCredentialsMultiFactor as UserCredentialsMultiFactorStorage;
+use Balloon\Hook;
+use Balloon\App\Idp\Hook\MultiFactorAuth;
+use Balloon\Server;
 
 return [
     OAuth2Server::class => [
         'arguments' => [
-            'storage' => '{'.MongoDB::class.'}',
+            'storage' => '{'.UserCredentialsStorage::class.'}',
             'config' => [
                 'enforce_state' => true,
                 'allow_implicit' => true,
                 'use_openid_connect' => true,
                 'issuer' => 'balloon'
             ],
-            'grant_types' => [
+            'grantTypes' => [
                 'user_credentials' => '{'.UserCredentials::class.'}',
+                'user_credentials_mfa' => '{'.UserCredentialsMultiFactor::class.'}',
                 'refresh_token' => '{'.RefreshToken::class.'}',
             ]
         ],
         'services' => [
+            /*UserCredentialsInterface::class => [
+                'use' => UserCredentialsStorage::class,
+                'merge' => false,
+            ],*/
+            RefreshTokenInterface::class => [
+                'use' => UserCredentialsStorage::class,
+            ],
             UserCredentials::class => [
                 'arguments' => [
-                    'storage' => '{'.MongoDB::class.'}'
+                    'storage' => '{'.UserCredentialsStorage::class.'}'
                 ]
+            ],
+            UserCredentialsMultiFactor::class => [
+                'arguments' => [
+                    'storage' => '{'.UserCredentialsMultiFactorStorage::class.'}'
+                ]
+            ],
+            UserCredentialsMultiFactorStorage::class => [
+                'merge' => false
+            ],
+            Server::class => [
+                'singleton' => false
             ]
         ]
     ],
@@ -45,6 +72,7 @@ return [
     Migration::class => [
         'calls' => [
             Installation::class => [
+
                 'method' => 'injectDelta',
                 'arguments' => ['delta' => '{'.Installation::class.'}']
             ],
@@ -55,6 +83,14 @@ return [
             'token' => [
                 'method' => 'injectAdapter',
                 'arguments' => ['adapter' => '{'.Token::class.'}', 'name' => 'token']
+            ],
+        ],
+    ],
+    Hook::class => [
+        'calls' => [
+            MultiFactorAuth::class => [
+                'method' => 'injectHook',
+                'arguments' => ['hook' => '{'.MultiFactorAuth::class.'}']
             ],
         ],
     ],